+++ core/src/ir/mod.rs
@@ -1,6 +1,14 @@
 //! Ligen intermediate representation.
 
-mod atomic;
+pub use attribute::*;
+pub use constant::*;
+pub use function::*;
+pub use identifier::*;
+pub use implementation::*;
+pub use literal::*;
+pub use parameter::*;
+pub use types::*;
+
 mod attribute;
 mod constant;
 mod function;
@@ -10,12 +18,5 @@ mod literal;
 mod parameter;
 mod types;
 
-pub use atomic::*;
-pub use attribute::*;
-pub use constant::*;
-pub use function::*;
-pub use identifier::*;
-pub use implementation::*;
-pub use literal::*;
-pub use parameter::*;
-pub use types::*;
+pub mod processing;
+
