+++ core/src/ir/processing/mod.rs
@@ -1,0 +1,5 @@
+//! Module containing traits for processing the IR.
+
+mod replace_identifier;
+
+pub use replace_identifier::*;
